First prototype of stateless protocol
In this first commit I'm sketching out how the stateless protocol works.
The server (Node.js) handles a connection by streaming a sequence of
numbers to the client (Ruby), which reads as many numbers as it wants
and prints their sum.

There is currently no error handling in the client -- this prototype is
purely to figure out if this design is workable. Particularly, whether
it deals nicely with the size of numbers involved.

The largest value this protocol should transmit is 0xff * 2 ^ 0xffff,
which is 65,543 bits in size. We use JavaScript's BigInt class to avoid
losing precision, and Ruby has baked-in big integers so that's not a
problem.

The serialisation could be a lot more efficient -- the numbers sent in
this example will always be some single-byte value times a power of 2,
so serialising large values in binary format will produce at most two
non-zero bytes and then many null bytes. We could use a format with much
better space efficiency, but the savings would not be so pronounced when
sending arbitrary numbers.

I've gone with line-separated JSON here as it performs acceptably well
on the size of number we're dealing with -- there is a little slow-down
on the larger numbers we'd expect, but not enough that the server gets
really bogged down serialising numbers.

This format lets me get a design off the ground quickly, without writing
lots of custom serialisation logic, especially as the JS BigInt class
can take decimal strings as input.

I may revisit the serialisation later if it proves problematic. For
example, we could invent a representation of the form (n, p)
representing n * 2^p. This would save space on the wire and remove the
need to use bigints on the server, as multiplying by 2 just means
incrementing p. The format would also affect how the reconnection state
is transmitted, so I'd like to get that done first before committing to
a format.