Extract application logic for the server
So that we can begin automated testing, and work out what state we need
to produce a resumable protocol, here we extract a Stateless class from
the server code. It can work in two modes:

- `Stateless.random()` returns a new stream starting at a random value
  in the range 1..0xff, emitting this first value

- `Stateless.resume(n)` returns a stream that resumes from the given
  value, but not emitting that value; the idea is that the client would
  provide this starting point from its last received message

The caller can pass a string from a JSON message directly into
Stateless.resume() as the BigInt constructor turns this into a number.
So the client can recover be simply sending the last data value it
received in an application-agnostic manner.