Extract Server class
In order to write tests for the client-server protocol, here we extract
Server from the server program. This lets us start a TCP connection to
the server and observe that it invokes the underlying app in the right
way, returning a stream of values in JSON format. The numbers are
strings so that they're safe to parse in a language where big integers
are a library feature.

Exposing Stateless.seed() as a static function means we can stub it out
to control the randomness for new clients.

This extraction also lets us configure the Server to control its timer
interval, so our tests aren't waiting a full second between messages as
the server would in production.