Make the server handle reconnections
Although the Stateless stream can be constructed to resume from a given
value, the Server did not expose this functionality. Here we make the
server read an initial message from the client, which is a JSON message
containing an optional state value. The client should pass the data
value of the last message it received here.

The server will detect this value and return a stream that picks up from
that value, rather than starting a new random stream.