Extract connection handling for the client
Having implemented a server protocol for resuming a stateless stream, we
can now implement a client to take advantage of this. The application
logic ("take N numbers from a stream and sum them") should ideally
remain isolated and not be complicated by connection-handling concerns.
Instead, an abstraction should exist that allows the app to reliably
fetch N messages.

The Client class introduced here acoomplishes this by exposing a method
for reading messages -- it does not care about the app's request pattern
or what the app will do with the messages. It attempts to read a message
from a TCP socket, opening one if necessary.

If it gets an EOFError or ECONNRESET while reading a message, that means
the connection was interrupted and we should immediately attempt to
reconnect.

If we get ECONNREFUSED it means we reached the server but it was not
listening on the specified port, or rejected our connection. If we get a
SocketError, that can indicate other failures to connect, e.g. a failure
to resolve a hostname using DNS -- we may be offline, or the host may
genuinely not exist, or there may be a transient DNS problem. In these
cases, we should back off as we're not currently able to connect and we
can try again later.

The backoff stategy here is a simple arithmetic delay that increases
with each attempt. This is sufficient to make the client not churn in a
loop trying to connect, while not resulting in huge delays as would
happen with an exponential strategy. Attempting to reconnect once per
minute at most seems reasonable, but we could easily change the policy
here without affecting the application.

To enable resuming the stream, Client maintains a cache of the last
value of the message.data field that it saw. On reconnection, it sends
this in its initial message to make the server resume from that point.
It only hands off the data field to the caller, so the application is
not exposed to the internal message structure used for reconnection
logic.