Refactor the client into classes
Now that we have a working client for the stateless protocol, we can
split out the various aspects of it into classes encapsulating the
various moving parts.

The Client class remains as a way to handle re-establishing a TCP
connection using the stateless protocol; it hides the reconnection logic
and returns the complete stream of values to the caller.

The Sum class takes a Client and sums the first N messages it receives;
this is the application built atop the stateless protocol.

The App class wraps up the logic of the command-line program, parsing
its arguments and returning the result of running the app, given a
stubbed-out client which we assume will return the complete message
stream.

The bin/client file is reduced to calling App#run and printing its
result.