Introduce the stateful server protocol
Here we add server-side support for a stateful protocol. It accepts an
initial request from the client containing its UUID and the number of
values it wants to receive. The server then streams a random sequence of
values out to the client, attaching the CRC-32 of the whole stream to
the final message. The messages have the form:

    { data: { value: Number, crc: Number? } }

There is currently no support for resuming the stream, so the Sessions
class only stores two pieces of state for each client: the running
CRC-32 of its messages, and the number of messages remaining to be sent.

The Sessions class doesn't care what is stored in the record for each
client -- it merely stores it and the Stateful class can do as it wants
with the stored data. The only important thing here is that the put()
method returns a Promise of a message that resolves after the state has
been persisted, so we only send messages whose effect has been
committed.