Split Client into two concerns
As we introduce support for the Stateful protocol, the
connection-handling splits into two concerns: how to establish a TCP
connection and detect it disconnecting, and what messages to send when a
client first connects or reconnects.

Here we put the former in the Client class -- this deals with making a
TCP connection, handling disconnects, sending initial messages and
reading JSON-structured data from the stream.

Client::Stateless wraps this in appropriate state for the stateless
protocol: tracking the last message received and re-sending it when a
connection is established.