Add reconnection support in the server
Here we extend the Sessions class, the Stateful app and the Handler
class to handle reconnection. We assign each new message a sequential ID
and persist the message before sending it to the client. The
Sessions.put() method is constructed so that the caller constructs the
next message and the new state from the existing state. The Sessions
class then assigns an ID to the message and returns it.

When a client wants to reconnect, it sends a 'state' value instead of
the 'params' value used on first connection. This tells the server the
ID of the last message the client received. It truncates its internal
queue to that point and begins replaying the messages the client does
not have. Once that process is complete, we begin generating new
messages that update the CRC state.