Implement stateful reconnection in the client
Using the protocol offered by the server, the client can now statefully
reconnect and receive the complete stream. Whereas its initial
connection message is of the form

    { uuid, params: { count } }

when it reconnects, it sends the message

    { uuid, state: id }

where 'id' is the maximum message ID it has seen so far. This message
informs the server that this is a reconnecting client and acknowledges
how much of the stream the client has seen, so that the server can
replay and missed messages before resuming.