Acknowledge received messages without replay
In order to avoid the server's message buffer growing indefinitely, the
client should periodically notify it that it has received a certain
message. This lets the server discard messages it's holding as the
client has confirmed receive of them.

Sending an ack message should not result in already-sent messages being
replayed, even if the ack is for an earlier message. An ack is purely
informational and does not indicate the client is reconnecting or has
missed later messages.