Integration test for the client and server
Here we add an integration test that the client is able to receive a
complete message stream from the server in the presence of connection
failures. The test starts the server process, then runs the client in
CRC mode and checks that it produces a successful result.

We could simulate failure by randomly killing the server process.
However, as all the client state is in memory, this is not workable with
the current design. Instead, we wrap the TCPSocket class that the client
uses in another class that randomly disconnects while reading and
writing messages, raising the same exceptions that TCPSocket does in
production.