Simplify session state and API
It is not strictly necessary for the Sessions class to persist the
"lag", the number of messages behind the most-recently-generated one the
client is at. That's because the process handling the connection has
access to the last ID the client saw, either because it was provided in
the client's initial message, or it was part of the last message the
server sent to the client.

Therefore it's sufficient for the server process to simply track the
last ID for the connected client, and each time it wants to send a
message to the client, it asks Sessions for the message after that last
ID. If no such message already exists, then one is generated, persisted
to storage, and sent to the client.