Make random number sequence deterministic
By replacing the JS Math.random() calls with a Mersenne twister, we can
make the message stream following the initial seed completely
deterministic but sitll pseudo-random from the client's point of view.
Storing the last random value in the state means the client will get the
same message sequence regardless of which server it connects to.